+++ engine/drizzle/0001_common_blackheart.sql
@@ -1,0 +1,3 @@
+ALTER TYPE "public"."user_role" ADD VALUE 'manager' BEFORE 'editor';--> statement-breakpoint
+ALTER TYPE "public"."user_role" ADD VALUE 'author';--> statement-breakpoint
+ALTER TYPE "public"."user_role" ADD VALUE 'reviewer';